Xen AI/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
torch
torchvision
pillow
numpy
